feat: 관리자 정보 수정 API 구현

+++ src/main/java/com/yeo_li/yeol_post/category/CategoryService.java
@@ -13,7 +13,7 @@ public class CategoryService {
     return categoryRepository.findCategoryByCategoryName(categoryName);
   }
 
-  public Category findCategoryByCategoryId(int categoryId) {
+  public Category findCategoryByCategoryId(Long categoryId) {
     return categoryRepository.findCategoryById(categoryId);
   }
 }
